% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{blankPlot}
\alias{blankPlot}
\title{Creates a blank plot canvas based on dimensions provided. Does not return anything.}
\usage{
blankPlot(...)
}
\description{
Creates a blank plot canvas based on dimensions provided. Does not return anything.
}
\examples{
blankPlot(x=c(0,10),y=c(0,20))

}
