% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{cleanNames}
\alias{cleanNames}
\title{Cleans up names of imported data frames}
\usage{
cleanNames(x)
}
\value{
data.frame
}
\description{
Cleans up names of imported data frames
}
\examples{
cleanNames(data)

}
