% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomkit_functions.R
\name{createMap}
\alias{createMap}
\title{Creates a mapping vector based on data.frame columns
supports piping a la dplyr}
\usage{
createMap(dat, base, name)
}
\value{
Vector
}
\description{
Creates a mapping vector based on data.frame columns
supports piping a la dplyr
}

