% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{exec_if}
\alias{exec_if}
\title{Execute If}
\usage{
exec_if(x, cond, .f)
}
\description{
This function executes a function upon an object if condition are met.
}
\examples{
x \%>\%
	exec_if(TRUE, ~mutate(.x, foo="bar"))


}
