% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{leadgr}
\alias{leadgr}
\title{String function that adds leading zeroes to digits.}
\usage{
leadgr(x, digits = 1)
}
\value{
String
}
\description{
String function that adds leading zeroes to digits.
}
\examples{
leadgr(42, 4)

}

