% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{pipeMarker}
\alias{pipeMarker}
\title{Prints something then passes object along
Text marker for long piped processes}
\usage{
pipeMarker(x, text = x)
}
\value{
Anything
}
\description{
Prints something then passes object along
Text marker for long piped processes
}

