% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pNum}
\alias{pNum}
\title{shortened version of prettyNum with diff defaults}
\usage{
pNum(x)
}
\value{
String
}
\description{
shortened version of prettyNum with diff defaults
}
\examples{
pNum(5000000)

}
