% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomkit_functions.R
\name{readkey}
\alias{readkey}
\title{The length of a string (in characters).}
\usage{
readkey()
}
\value{
None
}
\description{
The length of a string (in characters).
}
\examples{
... some code
readkey()

}

