% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{returnPart}
\alias{returnPart}
\title{functional way of returning specific part of data object}
\usage{
returnPart(x, y)
}
\value{
Any
}
\description{
functional way of returning specific part of data object
}
\examples{
returnPart(data, objectName)

}

