% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trimall}
\alias{trimall}
\title{Removes leading and trailing spaces.}
\usage{
trimall(tstring)
}
\value{
String
}
\description{
Removes leading and trailing spaces.
}
\examples{
trimall("  this is result   ")

}
