% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_functions.R
\name{exec_if}
\alias{exec_if}
\title{Calls an anonymous function if conditions are met}
\usage{
exec_if(x, cond, .f)
}
\value{
Anything
}
\description{
Calls an anonymous function if conditions are met
}
