% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{readkey}
\alias{readkey}
\title{Adds interactive pause to code.  Waits for user input.}
\usage{
readkey()
}
\value{
None
}
\description{
Adds interactive pause to code.  Waits for user input.
}
\examples{
... some code
readkey()

}

